requirements.txt

python-dotenv>=0.21.0
google-api-python-client>=2.97.0
langchain>=0.0.342
langchain-community>=0.0.1
requests>=2.28.0
beautifulsoup4>=4.11.2
chromadb>=0.4.21
openai>=0.27.0
streamlit>=1.20.0
pypdf2>=3.0.0
tiktoken>=0.4.0
faiss-cpu
pysqlite3-binary
lxml
tenacity
